Move Data and CUQIarray classes to a separate module (array module)
