+++ cuqi/array/__init__.py
@@ -1,0 +1,1 @@
+from ._array import CUQIarray
